hsa-runtime: finish Darwin build — libhsa-runtime64.dylib links
Takes the macOS ROCR build from 38/46 files compiling and no link, to
a working arm64 libhsa-runtime64.1.21.0.dylib with HSA public API
symbols (hsa_init, hsa_agent_iterate_*, hsa_queue_create, ...) exported.

New shim header:
- core/inc/link_darwin.h: Darwin-local stub for glibc's <link.h>.
  Defines the narrow subset of `struct link_map` and `struct r_debug`
  ROCR's loader publishes for debugger introspection (l_addr, l_name,
  l_prev/l_next, r_map, r_state + RT_CONSISTENT/RT_ADD/RT_DELETE).
  Darwin has no equivalent of glibc's _r_debug protocol; the structs
  compile and ROCR's internal bookkeeping still works, but a
  Darwin-aware debugger hook is future work.

Source edits (gated on __APPLE__):
- loader/executable.{hpp,cpp}, core/runtime/runtime.cpp,
  core/runtime/amd_topology.cpp: include link_darwin.h instead of
  <link.h>. The topology-side driver-discovery array is sized to 0
  and left empty on Darwin (no KFD/XDNA — MacOsDriver will populate
  it later).

- core/runtime/thunk_loader.cpp: skip the else-branch that takes direct
  addresses of ~99 hsaKmt* symbols, since we don't link libhsakmt on
  Darwin. Thunk table stays null; MacOsDriver path doesn't route
  through it.

- core/runtime/runtime.cpp::InteropMap: return HSA_STATUS_ERROR on
  Darwin (hsaKmtRegisterGraphicsHandleToNodesExt is KFD-only).

- core/runtime/runtime.cpp: four size_t vs uint64_t ABI fixes. Darwin
  arm64 has size_t == unsigned long and uint64_t == unsigned long
  long — distinct typedefs at identical width, so ExportDMABuf's
  size_t* out-param can't accept a uint64_t*; changed local vars to
  size_t. Also Min(len, importSize - fragOffset) now static_casts
  importSize to size_t.

- core/runtime/amd_gpu_agent.cpp: same fix — change local `available`
  from size_t to uint64_t to match core::Driver::AvailableMemory.

- core/runtime/amd_blit_kernel.cpp:
  * std::min(size, uint64_t(...)) ambiguous — cast explicitly.
  * The else branch for std::atomic_ref<uint32_t> (C++20) is gated to
    !(Linux||Darwin); add Darwin to the __atomic_store_n branch since
    it's a clang builtin and not C++20-only.

- libamdhsacode/amd_elf_image.cpp:
  * <sys/sendfile.h> → <copyfile.h> on Darwin; replace sendfile loop
    with fcopyfile (Darwin since 10.5, fd-to-fd).
  * brew libelf defines Elf64_Xword as `unsigned long` (not
    unsigned long long), so std::max / alignUp calls with a (uint64_t)1
    literal are ambiguous. Normalize all the ELF header-field
    comparisons to uint64_t via static_cast at the call site.
  * `edata->d_align` (libelf Elf_Data, size_t field) gets a matching
    size_t literal in std::max.

- libamdhsacode/amd_hsa_code_util.hpp: add size_t alignUp() overload
  so calls that pass two size_t args resolve unambiguously on Darwin
  arm64 (they would otherwise see uint64/uint32 overloads neither of
  which matches exactly).

- hsa-runtime CMakeLists.txt:
  * Skip Linux-only link flags on APPLE (ld64 rejects -Bdynamic,
    -z,noexecstack, --version-script, --enable-new-dtags). Also drop
    -lrt (Darwin's clock_gettime is in libc).
  * target_link_libraries APPLE branch: elf + dl + pthread only.

Verified: /tmp/rocr-darwin-build/rocr/lib/libhsa-runtime64.1.21.0.dylib
is a 4.2 MB Mach-O arm64 dylib with the expected HSA symbols.

Follow-up (out of scope for this commit):
- A Darwin-native HSA_ROCR debugger hook to replace the _r_debug
  protocol stub.
- MacOsDriver backend (Stage 2B): add DriverType::MACOS_DEXT,
  plug into the now-empty discover_driver_funcs array in
  amd_topology.cpp.
- Runtime load-time validation that loader URI lookup from mapped
  memory still works without /proc/self/maps and dl_iterate_phdr
  (paths currently fall through to the memory-URI fallback).

+++ projects/rocr-runtime/runtime/hsa-runtime/core/runtime/amd_gpu_agent.cpp
@@ -564,7 +564,10 @@ void GpuAgent::ReserveScratch()
     reserved_sz = MaxScratchDevice();
   }
 
-  size_t available;
+  // core::Driver::AvailableMemory takes uint64_t*; on Darwin arm64 size_t is
+  // unsigned long and uint64_t is unsigned long long — distinct typedefs
+  // even at identical width, so a pointer to size_t can't be passed.
+  uint64_t available;
   hsa_status_t err = driver().AvailableMemory(node_id(), &available);
   assert(err == HSA_STATUS_SUCCESS && "AvailableMemory failed");
   std::lock_guard<std::mutex> lock(scratch_lock_);
